requirements.txt
torch>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
statsmodels>=0.13.0
requests>=2.25.0
tqdm>=4.62.0
pyyaml>=6.0
tensorboard>=2.8.0
plotly>=5.0.0